Moved print to graphUtils
