+++ src/CMakeLists.txt
@@ -1,6 +1,7 @@
 target_sources(${MAIN_LIB} PRIVATE
   graphdbio.cpp
   graph.cpp
+  graphUtils.cpp
   graphdb.cpp
 )
